requirements.txt
amqp==2.6.1
asgiref==3.2.10
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
cssselect==1.1.0
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-appconf==1.0.4
django-celery-email==3.0.0
django-celery-results==1.2.1
django-environ==0.4.5
django-extensions==3.0.9
django-imagekit==4.0.2
django-quill-editor==0.1.16
django-social-share==2.0.0
django-taggit==1.3.0
djangorestframework==3.11.1
environ==1.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
Jinja2==2.11.2
kombu==4.6.11
lxml==4.5.2
markdown2==2.3.9
MarkupSafe==1.1.1
oauthlib==3.1.0
pilkit==2.0
Pillow==7.2.0
psycopg2-binary==2.8.5
pyquery==1.4.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
readtime==1.1.1
render==1.0.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.50.0
supervisor==4.2.1
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
